Rename do pacote
Renomeando o nome do pacote de steam_data para game_data para ficar mais genérico.
