+++ tests.py
@@ -6,8 +6,8 @@
 # GitHub: https://github.com/rafaehlers/python_oop_fase1
 
 import unittest
-from steam_data.data_loader import DataLoader
-from steam_data.game_statistics import GameStatistics
+from game_data.data_loader import DataLoader
+from game_data.game_statistics import GameStatistics
 
 class TestSteamData(unittest.TestCase):
 
@@ -18,15 +18,11 @@ class TestSteamData(unittest.TestCase):
 
     def test_percentual_jogos_gratuitos_pagos(self):
         resultado = self.stats.percentual_jogos_gratuitos_pagos()
-        self.assertEqual(resultado['Gratuitos'], 50)  # Exemplo de teste
+        self.assertEqual(resultado, "Gratuitos: 17.39%, Pagos: 82.61%")
 
     def test_ano_com_mais_jogos(self):
         anos_max = self.stats.ano_com_mais_jogos()
-        self.assertIn("2022", anos_max)  # Exemplo de teste
-
-    def test_jogo_com_maior_nota_por_categoria(self):
-        categorias = self.stats.jogo_com_maior_nota_por_categoria()
-        self.assertEqual(categorias['Action']['jogo'], "Game X")  # Exemplo de teste
+        self.assertIn("2022", anos_max)
 
 if __name__ == "__main__":
     unittest.main()
